FAS-43 FAS-44
breakout routes to outside of application entrypoint

pluralize entries in URI
