+++ app/main.py
@@ -1,12 +1,11 @@
-from fastapi import FastAPI, Depends, HTTPException, APIRouter
+from fastapi import FastAPI, Depends, APIRouter
 
-from typing import Any
 from sqlalchemy.orm import Session
 
-from app.schemas.todo_entry import TodoEntryCreate, TodoEntryUpdate, TodoEntry
+from app.schemas.todo_entry import TodoEntry
 from app import deps
 from app import crud
-
+from app.routers import todoentries
 
 # Initialize the app
 app = FastAPI(
@@ -14,12 +13,13 @@ app = FastAPI(
     description="A simple API for managing a todo list",
     openapi_url="/api/v1/openapi.json"
 )
+app.include_router(todoentries.router, prefix="/api/v1")
 
-# Initialize the router
+# Initialize the root route
 api_router = APIRouter()
 
 
-# Define the routes
+# Define the root route
 @api_router.get("/", status_code=200, response_model=dict[str, list[TodoEntry]])
 def root(
     db: Session = Depends(deps.get_db)
@@ -33,70 +33,7 @@ def root(
     return {"entries": entries}
 
 
-@api_router.get("/entry/{id}", status_code=200, response_model=TodoEntry)
-def fetch_todo_entry(
-    *,
-    id: int,
-    db: Session = Depends(deps.get_db),
-) -> Any:
-    """
-    Fetch a single to-do entry by id
-    """
-    result = crud.todo_entry.get(db=db, id=id)
-    if not result:
-        raise HTTPException(status_code=404,
-                            detail=f"Entry not found for id {id}")
-    return result
-
-
-@api_router.post("/entry", status_code=201, response_model=TodoEntry)
-def create_entry(
-    *,
-    entry_in: TodoEntryCreate,
-    db: Session = Depends(deps.get_db)
-) -> dict:
-    """
-    Create a new to-do entry
-    """
-    entry = crud.todo_entry.create(db=db, obj_in=entry_in)
-    return entry
-
-
-@api_router.put("/entry", status_code=200, response_model=TodoEntry)
-def update_entry(
-    *,
-    entry_in: TodoEntryUpdate,
-    db: Session = Depends(deps.get_db)
-) -> Any:
-    """
-    Update an existing to-do entry
-    """
-    entry = crud.todo_entry.get(db=db, id=entry_in.id)
-    if not entry:
-        raise HTTPException(status_code=404,
-                            detail=f"Entry not found for id {entry_in.id}")
-    entry = crud.todo_entry.update(db=db, db_obj=entry, obj_in=entry_in)
-    return entry
-
-
-@api_router.delete("/entry/{id}", status_code=200, response_model=TodoEntry)
-def delete_entry(
-    *,
-    id: int,
-    db: Session = Depends(deps.get_db)
-) -> Any:
-    """
-    Delete an existing to-do entry
-    """
-    entry = crud.todo_entry.get(db=db, id=id)
-    if not entry:
-        raise HTTPException(status_code=404,
-                            detail=f"Entry not found for key {id}")
-    entry = crud.todo_entry.remove(db=db, id=entry.id)
-    return entry
-
-
-app.include_router(api_router)
+app.include_router(todoentries.router)
 
 
 if __name__ == "__main__":
