+++ tests/conftest.py
@@ -11,10 +11,7 @@ from sqlalchemy.orm import sessionmaker
 from app.db.base import Base
 from app.deps import get_db
 from app.main import api_router
-
-# import sys
-# import os
-# sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
+from app.routers import todoentries
 
 
 SQLALCHEMY_DATABASE_URL = "sqlite:///tests/config/test.db"
@@ -27,6 +24,8 @@ SessionTesting = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 def start_test_app():
     app = FastAPI()
     app.include_router(api_router)
+    # below can be broken out for more granular testing, e.g. mservice architecture
+    app.include_router(todoentries.router)
     return app
